Refactor: Improve API, dynamic imports, and telemetry
This commit introduces several improvements:
- Refactors the OpenAI API client to handle missing API keys gracefully.
- Sets `dynamic = "force-dynamic"` for several pages to ensure they fetch fresh data.
- Introduces `Suspense` for dynamic imports in `FenLabyrinthPage` and `NpcPromptCommandPage` for better loading states.
- Moves the `GalaxyHubClient` component to its own file and uses `next/dynamic` for server-side rendering prevention.
- Updates `tsconfig.json` for better module resolution and target compatibility.
- Refactors telemetry to use an async queue and adds new event metadata.
- Moves `CONSUMERS` definition to a dedicated `lib/trinity-consumers.ts` file.
- Filters `listWorlds` to only include renderable 3DT destinations.
- Adds a new `app/intake/page.tsx` route.
