+++ app/chaos-vault/io/page.tsx
@@ -1,6 +1,8 @@
 import { WcmLink } from "../../../components/wcm/WcmLink";
 import { WcmNavPanel } from "../../../components/wcm/WcmNavPanel";
 
+export const dynamic = "force-dynamic"
+
 export default function ChaosVaultIoPage() {
   return (
     <main>
