+++ app/intake/page.tsx
@@ -1,0 +1,8 @@
+"use client"
+
+import IntakeProtocolPanel from "@/components/IntakeProtocolPanel"
+
+export default function IntakePage() {
+  return <IntakeProtocolPanel />
+}
+
