+++ app/world/clear/layout.tsx
@@ -1,5 +1,6 @@
 import type { ReactNode } from "react"
-import { TrinityProvider, CONSUMERS } from "@/components/trinity-provider"
+import { TrinityProvider } from "@/components/trinity-provider"
+import { CONSUMERS } from "@/lib/trinity-consumers"
 
 export const metadata = {
   title: "CLEAR | WIRED CHAOS",
