+++ app/world/signal-duel/page.tsx
@@ -1,5 +1,7 @@
 "use client"
 
+export const dynamic = "force-dynamic"
+
 import type React from "react"
 
 import { useState, useEffect, useCallback } from "react"
